Delete Seed's BackupBucket only if no BackupEntries are referencing it

+++ pkg/gardenlet/controller/seed/seed/reconciler_delete.go
@@ -73,18 +73,28 @@ func (r *Reconciler) delete(
 		return reconcile.Result{}, nil
 	}
 
-	if seed.Spec.Backup != nil {
-		backupBucket := &gardencorev1beta1.BackupBucket{ObjectMeta: metav1.ObjectMeta{Name: string(seed.UID)}}
-		if err := r.GardenClient.Delete(ctx, backupBucket); client.IgnoreNotFound(err) != nil {
-			return reconcile.Result{}, err
-		}
-	}
-
 	// Before deletion, it has to be ensured that no Shoots nor BackupBuckets depend on the Seed anymore.
 	// When this happens the controller will remove the finalizers from the Seed so that it can be garbage collected.
 	associatedShoots, err := controllerutils.DetermineShootsAssociatedTo(ctx, r.GardenClient, seed)
 	if err != nil {
 		return reconcile.Result{}, err
+	}
+
+	if seed.Spec.Backup != nil {
+		backupBucket := &gardencorev1beta1.BackupBucket{ObjectMeta: metav1.ObjectMeta{Name: string(seed.UID)}}
+		associatedBackupEntries, err := controllerutils.DetermineBackupEntryAssociations(ctx, r.GardenClient, backupBucket.Name)
+		if err != nil {
+			return reconcile.Result{}, err
+		}
+		if len(associatedBackupEntries) > 0 {
+			log.Info("Cannot delete Seed's BackupBucket because BackupEntries are still referencing it", "backupEntryNames", associatedBackupEntries)
+			r.Recorder.Eventf(backupBucket, corev1.EventTypeNormal, v1beta1constants.EventResourceReferenced, "cannot delete BackupBucket because the following BackupEntries are still referencing it: %+v", associatedBackupEntries)
+			return reconcile.Result{}, fmt.Errorf("BackupBucket %s still has references", backupBucket.Name)
+		}
+
+		if err := r.GardenClient.Delete(ctx, backupBucket); client.IgnoreNotFound(err) != nil {
+			return reconcile.Result{}, err
+		}
 	}
 
 	associatedBackupBuckets, err := controllerutils.DetermineBackupBucketAssociations(ctx, r.GardenClient, seed.Name)
